+++ src/controllers/admin/film/Series/index.js
@@ -18,10 +18,13 @@ exports.postCreateSeries = AsyncHandler(async (req, res, next) => {
 
   const series = await Series.create({
     title: req.body.title,
+    releaseDate: req.body.releaseDate,
     description: req.body.description,
+    director: req.body.director,
+    cast: req.body.cast,
+    country: req.body.country,
     imageUrl: infoImage,
-    listSeriesId:
-      req.body.listSeriesId !== 'none' ? req.body.listSeriesId.split(',') : [],
+    listCategoryId: req.body.listCategoryId.split(','),
     createAt: Date.now(),
     createBy: '6543c28ae4b2dbdf546106c3',
   });
@@ -45,25 +48,31 @@ exports.postCreateSeries = AsyncHandler(async (req, res, next) => {
 
 exports.postDeleteSeries = AsyncHandler(async (req, res, next) => {
   console.log(req.params);
+  console.log(req.body);
   if (!req.params.seriesId) {
     return next(
       new ErrorResponse(`Please enter a valid id series delete`, 404),
     );
   }
 
-  const series = await Series.findOne({ _id: req.params.seriesId });
+  if (!req.body.type) {
+    return next(new ErrorResponse(`Please send type series delete`, 404));
+  }
+
+  const series = await Series.findById(req.params.seriesId);
 
   if (!series) {
     return next(
-      new ErrorResponse(
-        `The system is experiencing problems, please try again later!!`,
-        401,
-      ),
+      new ErrorResponse(`Cannot find series id ${req.params.seriesId}!!`, 401),
     );
   }
-  await deleteImageCloud(series.imageUrl.imageId);
-
-  await Series.deleteOne({ _id: req.params.seriesId });
+  if (req.body.type === 'delete') {
+    series.isDelete = true;
+    await series.save();
+  } else {
+    await deleteImageCloud(series.imageUrl.imageId);
+    await Series.deleteOne({ _id: req.params.seriesId });
+  }
 
   res.status(201).json({
     success: true,
@@ -88,10 +97,13 @@ exports.postUpdateSeries = AsyncHandler(async (req, res, next) => {
   };
 
   series.title = req.body.title;
+  series.releaseDate = req.body.releaseDate;
   series.description = req.body.description;
+  series.director = req.body.director;
+  series.cast = req.body.cast;
+  series.country = req.body.country;
   series.imageUrl = infoImage;
-  series.listSeriesId =
-    req.body.listSeriesId !== 'none' ? req.body.listSeriesId.split(',') : [];
+  series.listCategoryId = req.body.listCategoryId.split(',');
   series.updateAt = Date.now();
   await series.save();
 
@@ -142,4 +154,20 @@ exports.posAddManySeries = AsyncHandler(async (req, res, next) => {
 
   console.log('end');
   await DeleteFile(req.file.path);
+});
+
+exports.postRecoverSeries = AsyncHandler(async (req, res, next) => {
+  const series = await Series.findById(req.body.dataId);
+
+  if (!series) {
+    return next(
+      new ErrorResponse(`Cannot find series id ${req.body.dataId}!!`, 401),
+    );
+  }
+  series.isDelete = false;
+  await series.save();
+  res.status(200).json({
+    success: true,
+    message: 'Changed the status series successfully',
+  });
 });
